Correct dev branch after the fallout from merging new TerminalInterface refactor.
