+++ src/main/java/mars/algorithm/limited/AlgorithmLimitedScopeAStar.java
@@ -1,5 +1,6 @@
-package mars.algorithm;
+package mars.algorithm.limited;
 
+import mars.algorithm.Algorithm;
 import mars.coordinate.AStarCoordinate;
 import mars.coordinate.Coordinate;
 import mars.out.MapImageOutput;
@@ -95,8 +96,7 @@ public class AlgorithmLimitedScopeAStar extends Algorithm {
             thisCoord = coords.get(coords.size()-1); //set current location to the latest position in the path
             //System.out.println((thisCoord.getX()) + "," + (thisCoord.getY())); //debug
         }
-        output = new TerminalOutput(path); //if we reached here, we got out of the while loop. we're done!
-        output = new MapImageOutput(path, map.getMapPath());
+        //If we reached here, we got out of the while loop. We're done!
     }
